refactor(runtime): extract web_augment module from agent_loop
The web-search augmentation cluster — `should_augment_web_search`,
`SEARCH_QUERY_GEN_PROMPT`, `generate_search_queries`, and
`web_search_augment` — moves into `agent_loop/web_augment.rs`.

Only `web_search_augment` is consumed from the main loop (called once
each at the head of `run_agent_loop` and `run_agent_loop_streaming`
before the first LLM call), so it is the only `use self::web_augment::*`
re-import at file scope. `should_augment_web_search` and
`SEARCH_QUERY_GEN_PROMPT` are referenced only from the `#[cfg(test)] mod
tests` block, so they are pulled in alongside the existing
`tool_resolution` test-only imports — same pattern as the prior commit,
no `#[allow(unused_imports)]` needed.

`generate_search_queries` stays module-private — it has no caller outside
this cluster. The new module is named `web_augment` (not `web_search`)
to avoid name-collision confusion with the sibling `crate::web_search`
module that holds the `WebToolsContext` it consumes.

Behavior-preserving: verbatim function bodies, only visibility was
adjusted (`pub(super)` on the 3 cluster-public items, private on
`generate_search_queries`). The Auto-mode fallback (assume tools
supported when `model_supports_tools` metadata is absent) and the
JSON-extraction sentinel-vs-error contract on `generate_search_queries`
are unchanged.

mod.rs: 13,531 -> 13,349 LOC (-182). cargo check + cargo clippy --lib
both clean.

Refs #3710
